pkg: add deduplicating url queue package
when we scrape a web page, we are going to

* parse the html
* find all the links (ones which are properly formed hrefs)
* parse them relative to the current URL
* go and scrape them, if we have not scraped them already

because we are using a language with good concurrency primitives, the
queueing part is fairly easy. however we should provide a nice interface
because there are some sharp edges with concurrency

1. we want to make our queue usable in parallel
2. for it to be usable in parallel, we need to make sure that we can
   terminate when the queue is empty
3. we can use a WaitGroup for this, but we should not expose it because
   it could be misused

additionally, we want to keep track of which URLs we have already
scraped, otherwise we may scrape the same page forever. keeping track of
the pages we have already scraped consumes a lot of memory if we record
all the URLs

instead of keeping a log of all of the URLs we can use a bloom filter. a
bloom filter probablistically keeps track of items we have already seen,
using some hashing

a bloom filter is tunable, you can choose what % of false positives you
want, which determines how much memory it uses

a bloom filter will never report a false negative, therefore if we "see"
a URL in our bloom filter, we have probably scraped it, but we will
never scrape a URL we have already scraped, which means we will avoid
loops

Signed-off-by: toby lorne <toby@toby.codes>